Create bot file and rename dev script to local
