+++ lib/commands.js
@@ -44,6 +44,6 @@ export const registerCommands = async (getApp, locales) => {
   await getApp().commands.put({ data })
 }
 
-export const deletelCommand = async (getApp, command) => {
+export const deleteCommand = async (getApp, command) => {
   await getApp().commands(command.id).delete()
 }
